Reorganize file structure in pom-analyzer project
