+++ analyzer/pom-analyzer/src/main/java/eu/fasten/analyzer/pomanalyzer/utils/EffectiveModelBuilder.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package eu.fasten.analyzer.pomanalyzer;
+package eu.fasten.analyzer.pomanalyzer.utils;
 
 import java.io.File;
