+++ analyzer/pom-analyzer/src/main/java/eu/fasten/analyzer/pomanalyzer/utils/MavenRepositoryUtils.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package eu.fasten.analyzer.pomanalyzer;
+package eu.fasten.analyzer.pomanalyzer.utils;
 
 import static eu.fasten.core.utils.Asserts.assertNotNull;
 import static java.io.File.createTempFile;
